pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "audio-text"
version = "0.1.0"
description = "Get text from audio"
authors = [{ name = "Daniel Ducuara", email = "daniel14015@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12.0"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]

dependencies = [
  # Core API dependencies (needed by both API and Worker)
  "fastapi==0.116.2",
  "uvicorn[standard]==0.35.0",
  "apischema==0.19.0",
  "python-magic==0.4.27",
  "pydantic[email]==2.12.4",
  # Storage and async processing
  "boto3==1.34.0",
  "google-cloud-pubsub>=2.21.0",
  # Database
  "sqlalchemy==2.0.0",
  "alembic==1.13.0",
  "asyncpg>=0.29.0",
  "psycopg2-binary==2.9.9",
  # Email
  "sendgrid==6.11.0",
  # Additional utilities
  "python-multipart==0.0.20",
  # "pydantic-settings==2.1.0",
]

[project.optional-dependencies]
# ML/Worker dependencies (only for Celery workers doing audio processing)
worker = ["faster-whisper>=1.0.0", "setuptools-rust==1.9.0", "requests==2.31.0"]

[dependency-groups]
lint = ["ruff>=0.12.10", "pre-commit>=4.3.0"]
test = [
  "pytest==6.2.4",
  "pytest-mock==3.6.1",
  "pytest-cov==2.12.1",
  "pytest-asyncio==0.15.1",
]

[tool.hatch.build.targets.wheel]
packages = ["audio_text_backend"]

[tool.hatch.envs.hatch-test]
default-args = ["tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=app --cov-report=term-missing"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
preview = true

[tool.ruff.lint]
extend-select = ["I", "D"]
fixable = ["I", "D", "F401"]
ignore = ["D1"]

[tool.ruff]
line-length = 100
# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
]
